+++ src/App.jsx
@@ -9,13 +9,12 @@ import Links from './pages/Links/Components/links'
 import Leadership from './pages/Leadership/Components/leadership'
 import Archives from './pages/Archives/Components/archives'
 import LeadershipBio from './pages/Leadership/Components/leadershipBio'
-import RootRouteComponent from './root'
 import Merch from './pages/Merch/Components/Merch';
-import Admin from './pages/Admin/Components/AdminLayout';
-import AdminLogin from './pages/Admin/Components/AdminLogin';
+import AdminLogin from './pages/Admin/AdminLogin/AdminLogin';
 import { AuthLayout } from './components/AuthLayout';
-import AdminDashboard from './pages/Admin/Components/AdminDashboard';
-import AdminLayout from './pages/Admin/Components/AdminLayout';
+import AdminDashboard from './pages/Admin/AdminDashboard/AdminDashboard';
+import AdminLayout from './pages/Admin/AdminLayout';
+import PageLayout from './pages/PageLayout';
 
 /* If adding new pages, add a new route with a relative link pointing to the new page and 
 ** set the element to the main class of your page 
@@ -26,7 +25,7 @@ import AdminLayout from './pages/Admin/Components/AdminLayout';
 const newRouter = createBrowserRouter(
     createRoutesFromElements(
         <Route element={<AuthLayout/>}>
-            <Route path='/' element={<RootRouteComponent />}>
+            <Route path='/' element={<PageLayout/>}>
                 <Route index element={<Home/>}/>
                 <Route path='about' element={<About/>}/>
                 <Route path='events' element={<Events/>}/>
@@ -36,10 +35,10 @@ const newRouter = createBrowserRouter(
                 <Route path='archives' element={<Archives/>}/>
                 <Route path='bios' element={<LeadershipBio/>}/>
             </Route>
-            
+
             <Route path='admin/login' element={<AdminLogin/>}/>
-            <Route path='admin' element={<Admin/>}/>
-            <Route element={<AdminLayout/>}>
+            <Route path='admin' element={<AdminLayout/>}>
+                <Route index element={<AdminDashboard/>}/> 
                 <Route path='dashboard' element={<AdminDashboard/>}/>  
             </Route>   
         </Route>
@@ -49,7 +48,7 @@ const newRouter = createBrowserRouter(
 const router = createBrowserRouter([
     {
         path: '/',
-        element: <RootRouteComponent />, /**RootRouteComponent is the parent of all the other Routes and will contain all shared Components (ex. 
+        element: <PageLayout />, /**PageLayout is the parent of all the other Routes and will contain all shared Components (ex. 
         Navbar, Footer, ScrollToTop) */
         children: [
             {
@@ -88,7 +87,7 @@ const router = createBrowserRouter([
     },
     {
         path: 'admin',
-        element: <Admin/>,
+        element: <AdminLayout/>,
         children: [
             {
                 path: 'login',
