--- src/Pages/Admin/Components/AdminLayout.jsx
@@ -1,15 +1,0 @@
-import { Outlet } from "react-router-dom";
-import { useAuth } from "../../../hooks/useAuth";
-import { Navigate } from "react-router-dom";
-
-export default function AdminLayout() {
-    const { user } = useAuth();
-
-    if (!user) {
-        return <Navigate to="/admin/login"/>;
-    }
-    
-    return(            
-        <Outlet/>
-    );
-}
